requirements.txt
a2wsgi==1.10.10
aiomysql==0.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
autopep8==2.3.2
bcrypt==4.3.0
black==22.1.0
blue==0.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
crcmod==1.7
cryptography==45.0.6
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
Faker==37.6.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastapi-sessions==0.3.2
fastsession==0.3.0
flake8==4.0.1
fonttools==4.59.2
fpdf==1.7.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
kiwisolver==1.4.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.6.1
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.3.2
orjson==3.11.1
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pyasn1==0.6.1
pycodestyle==2.8.0
pycparser==2.22
pydantic==2.11.7
pydantic-br==1.1.0
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==2.4.0
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.2.3
pypika-tortoise==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.2
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlmodel==0.0.24
starlette==0.47.2
tomli==2.3.0
tortoise-orm==0.25.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
validate_docbr==1.11.1
watchfiles==1.1.0
websockets==15.0.1
